Support for multi-threaded access in Agenda class
Currently certain functions which modify agenda UI element may be accesed from
thread which did not create this UI. This leads to .NET exception as such access
is not allowed. This commit modifies certain Add* routines to use .NET invoker
to allow such access in safe manner. Note that this support is not
complete yet. There is still some functionality such as Update/Remove in
Agenda which may be acccessedby different thread.
Also removing certain event handler when agenda UI is disposed, so that it is
not modified after disposing.